backend-python/requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx
boto3
google-auth
